requirements.txt
Django>=4.2
Pillow
django-ckeditor